Isolate `pr merge` command
